+++ exasol/toolbox/nox/tasks.py
@@ -1,6 +1,9 @@
 from __future__ import annotations
 
-from exasol.toolbox.cli import version
+from exasol.toolbox.nox._shared import (
+    Mode,
+    _version,
+)
 
 __all__ = [
     "Mode",
@@ -14,16 +17,13 @@ __all__ = [
     "build_docs",
     "open_docs",
     "clean_docs",
+    "prepare_release",
 ]
 
 import argparse
 import shutil
 import webbrowser
 from collections import ChainMap
-from enum import (
-    Enum,
-    auto,
-)
 from functools import partial
 from pathlib import Path
 from typing import (
@@ -41,13 +41,6 @@ from exasol.toolbox.metrics import (
     format_report,
 )
 from exasol.toolbox.project import python_files as _python_files
-from exasol.toolbox.release import (
-    Version,
-    extract_release_notes,
-    new_changelog,
-    new_changes,
-    new_unreleased,
-)
 from noxconfig import (
     PROJECT_CONFIG,
     Config,
@@ -57,11 +50,6 @@ _DOCS_OUTPUT_DIR = ".html-documentation"
 _PATH_FILTER = tuple(["dist", ".eggs", "venv"] + list(Config.path_filters))
 
 python_files = partial(_python_files, path_filters=_PATH_FILTER)
-
-
-class Mode(Enum):
-    Fix = auto()
-    Check = auto()
 
 
 def _context(session: Session, **kwargs: Any) -> MutableMapping[str, Any]:
@@ -100,12 +88,6 @@ def _pyupgrade(session: Session, files: Iterable[str]) -> None:
         "--exit-zero-even-if-changed",
         *files,
     )
-
-
-def _version(session: Session, mode: Mode, version_file: Path) -> None:
-    command = ["poetry", "run", "version-check"]
-    command = command if mode == Mode.Check else command + ["--fix"]
-    session.run(*command, f"{version_file}")
 
 
 def _pylint(session: Session, files: Iterable[str]) -> None:
@@ -334,119 +316,7 @@ def report(session: Session) -> None:
     print(format_report(project_report, fmt))
 
 
-@nox.session(name="prepare-release", python=False)
-def prepare_release(session: Session, python=False) -> None:
-    """
-    Prepares the project for a new release.
-    """
-
-    def _parser():
-        parser = argparse.ArgumentParser(
-            prog=f"nox -s prepare-release",
-            usage="nox -s prepare-release -- [-h] version",
-            formatter_class=argparse.ArgumentDefaultsHelpFormatter,
-        )
-        parser.add_argument(
-            "version",
-            type=version,
-            help=("A version string of the following format:" '"NUMBER.NUMBER.NUMBER"'),
-        )
-        parser.add_argument(
-            "--no-add",
-            default=False,
-            action="store_true",
-            help=("Neither add nor commit the changes"),
-        )
-        parser.add_argument(
-            "--no-branch",
-            default=False,
-            action="store_true",
-            help=("Do not create a branch to commit the changes on"),
-        )
-        parser.add_argument(
-            "--no-pr",
-            default=False,
-            action="store_true",
-            help=("Do not create a pull request for the changes"),
-        )
-        return parser
-
-    parser = _parser()
-    args = parser.parse_args(session.posargs)
-    new_version = args.version
-    old_version = Version.from_poetry()
-    if not (new_version >= old_version):
-        error_msg = (
-            f"Invalid version: the release version ({new_version}) "
-            f"must be greater than or equal to the current version ({old_version})."
-        )
-        session.error(error_msg.format(new=new_version, old=old_version))
-
-    if not args.no_branch and not args.no_add:
-        # prepare branch
-        session.run("git", "switch", "-c", f"release/prepare-{new_version}")
-
-    # bump project version and sync version file
-    session.run("poetry", "version", f"{new_version}")
-    _version(session, Mode.Fix, PROJECT_CONFIG.version_file)
-
-    # create a changelog file for the release and also create a new empty unrleased file
-    unreleased = Path(PROJECT_CONFIG.root) / "doc" / "changes" / "unreleased.md"
-    changelog = (
-        Path(PROJECT_CONFIG.root) / "doc" / "changes" / f"changes_{new_version}.md"
-    )
-    changes = Path(PROJECT_CONFIG.root) / "doc" / "changes" / f"changelog.md"
-
-    changelog_content = extract_release_notes(unreleased)
-    changelog.write_text(new_changelog(new_version, changelog_content))
-
-    unreleased.write_text(new_unreleased())
-
-    changes_content = new_changes(changes, new_version)
-    changes.write_text(changes_content)
-
-    if args.no_add:
-        return
-
-    # 3. commit changes
-    session.run("git", "add", f"{changelog}")
-    session.run("git", "add", f"{unreleased}")
-    session.run("git", "add", f"{changes}")
-    session.run("git", "add", f"{PROJECT_CONFIG.root / 'pyproject.toml'}")
-    session.run("git", "add", f"{PROJECT_CONFIG.version_file}")
-    session.run("git", "commit", "-m", f"Prepare release {new_version}")
-
-    # 4. create pr
-    if not args.no_pr:
-        session.run(
-            "gh",
-            "pr",
-            "create",
-            "--title",
-            f"Prepare release {new_version}",
-            "--body",
-            '""',
-        )
-
-
-@nox.session(name="release", python=False)
-def release(session: Session, python=False) -> None:
-    """
-    Creates a new release and publishing it to GitHub and pypi.
-    """
-    session.error("Not implemented yet")
-    # Precondition(s):
-    # Convert ci-cd.yml workflow to cd.yml workflow
-    # Tests validation can be skipped. Branch protection together with
-    # PR and merge validation shoudl be sufficient
-    # ----------------------------------------------------------------------
-    # 0. Check that version isn't released yet (tag does not exist (origin))
-    #   0.1. update git information
-    #   0.2. check if origin does not have the tag yet
-    # 1. check if current branch is main/master
-    # 2. build wheel/package
-    # 3. create release tag
-    # 4. push release tag to origin
-    # 5. publish on gh
-    # 5. publish on pypi
-    # 6. output relase message/information
+# Note:
+#   Nox orders the targets based on their definition, so in order to have the
+#   listing obey our disired ordering we need to introduce these defintions later on.
+from exasol.toolbox.nox._release import prepare_release  # pylint: disable=E402
